Add :isValid function to all custom objects
Since all methods are a no-op when objects are invalid, it's nice to
have a method that allows an E2 to query whether or not an object is
valid, for example when retrieving it from a table or a persist
variable.

The new method's documentation is as follows:

"Returns whether or not the element is valid. Elements that were not
created by the element's constructor, such as persist variables that
have not been assigned to, and table lookups that are not present, are
not valid and do not perform any action when modified."

Example code that was used to verify the method's functionality is as
follows:

```
@name
@inputs
@outputs
@persist Frame:dframe
@trigger

function void log(String: string) {
	print(_HUD_PRINTTALK, String)
}

function void logValidity(Frame: dframe) {
	Truthy = 0
	if (Frame) { Truthy = 1 }

	log("truthy: " + toString(Truthy) + ", valid: " + toString(Frame:isValid()))
}

log("Uninitialized persist (should be 0):")
logValidity(Frame)

log("Uninitialized table field (should also be 0):")
logValidity(table()["Frame",dframe])

log("Constructor function (should be 1):")
logValidity(dframe(0))
```